tarot_bot/requirements.txt
python-telegram-bot==21.10
requests==2.32.3
python-dotenv==1.0.1
openai==1.59.6
